--- app/src/main/java/com/example/cookbook/buscador/searching.values.kt
@@ -1,2 +1,0 @@
-package com.example.cookbook.buscador
-
