+++ app/src/main/java/com/example/cookbook/presentation/buscador/InitialSearch.kt
@@ -1,7 +1,6 @@
-package com.example.cookbook.buscador
+package com.example.cookbook.presentation.buscador
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
@@ -20,15 +19,10 @@ import androidx.navigation.NavController
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color.Companion.Gray
-import androidx.compose.ui.graphics.Color.Companion.Transparent
-import androidx.compose.ui.graphics.Color.Companion.White
-import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.navigation.compose.rememberNavController
 import com.example.cookbook.R
-import com.example.cookbook.myRecipe.MyRecipeView
 
 @Composable
 fun BuscadorinicialView(navController: NavController) {
