+++ app/src/main/java/com/example/cookbook/presentation/login/views/ChangePasswordView.kt
@@ -1,4 +1,4 @@
-package com.example.cookbook.login
+package com.example.cookbook.presentation.login.views
 
 import androidx.compose.foundation.BorderStroke
 import com.example.cookbook.R
@@ -20,10 +20,12 @@ import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.runtime.*
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.tooling.preview.Preview
 
 @Composable
-fun ForgotPasswordView() {
+fun ChangePasswordView() {
     Column(
         modifier = Modifier
             .fillMaxWidth()
@@ -63,22 +65,23 @@ fun ForgotPasswordView() {
 
         Spacer(modifier = Modifier.height(16.dp))
 
-        // Username Input
-        var username by remember { mutableStateOf(TextFieldValue("")) }
+        var password by remember { mutableStateOf(TextFieldValue("")) }
+        var passwordVisible by remember { mutableStateOf(false) }
         OutlinedTextField(
-            value = username,
-            onValueChange = { username = it },
-            label = { Text(text = "Username") },
+            value = password,
+            onValueChange = { password = it },
+            label = { Text(text = "New Password") },
+            visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
             modifier = Modifier.fillMaxWidth()
         )
 
         Spacer(modifier = Modifier.height(8.dp))
 
-        var email by remember { mutableStateOf(TextFieldValue("")) }
+        var confirm by remember { mutableStateOf(TextFieldValue("")) }
         OutlinedTextField(
-            value = email,
-            onValueChange = { email = it },
-            label = { Text(text = "Email") },
+            value = confirm,
+            onValueChange = { confirm = it },
+            label = { Text(text = "Confirm New Password") },
             modifier = Modifier.fillMaxWidth()
         )
 
@@ -96,7 +99,7 @@ fun ForgotPasswordView() {
             shape = RoundedCornerShape(30.dp),
             colors = ButtonDefaults.buttonColors(Color(0xFFFFFFFF))
         ) {
-            Text(text = "Send Verification Code", fontSize = 18.sp, color = Color(0xFFFFA500))
+            Text(text = "Change Password", fontSize = 18.sp, color = Color(0xFFFFA500))
         }
 
         Spacer(modifier = Modifier.height(50.dp))
@@ -105,6 +108,6 @@ fun ForgotPasswordView() {
 
 @Preview(showBackground = true)
 @Composable
-fun PreviewForgotPasswordView() {
-    ForgotPasswordView()
+fun PreviewChangePasswordView() {
+    ChangePasswordView()
 }
