+++ app/src/main/java/com/example/cookbook/presentation/myRecipe/MyRecipeView.kt
@@ -1,4 +1,4 @@
-package com.example.cookbook.myRecipe
+package com.example.cookbook.presentation.myRecipe
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
@@ -21,7 +21,6 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.rememberNavController
 import com.example.cookbook.R
