+++ app/src/main/java/com/example/cookbook/presentation/signup/SignupView.kt
@@ -1,4 +1,4 @@
-package com.example.cookbook.login
+package com.example.cookbook.presentation.signup
 
 import androidx.compose.foundation.BorderStroke
 import com.example.cookbook.R
@@ -14,18 +14,20 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.runtime.*
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.tooling.preview.Preview
+import androidx.navigation.NavController
+import androidx.navigation.compose.rememberNavController
 
 @Composable
-fun ChangePasswordView() {
+fun SignupView(navController: NavController) {
     Column(
         modifier = Modifier
             .fillMaxWidth()
@@ -42,7 +44,7 @@ fun ChangePasswordView() {
                 .align(Alignment.Start)
         )
 
-        Spacer(modifier = Modifier.height(70.dp))
+        //Spacer(modifier = Modifier.height(70.dp))
 
         //Image Logo
         Image(
@@ -54,40 +56,74 @@ fun ChangePasswordView() {
             contentScale = ContentScale.Crop
         )
 
-        Spacer(modifier = Modifier.height(40.dp))
+        //Spacer(modifier = Modifier.height(40.dp))
 
-        // Title Text
+// Title Text
         Text(
-            text = "Forgot your password?",
+            text = "Sign up",
             fontSize = 28.sp,
             fontWeight = FontWeight.Bold
         )
 
-        Spacer(modifier = Modifier.height(16.dp))
 
+        //Spacer(modifier = Modifier.height(16.dp))
+
+
+// Email Input
+        var email by remember { mutableStateOf(TextFieldValue("")) }
+        OutlinedTextField(
+            value = email,
+            onValueChange = { email = it },
+            label = { Text(text = "Email:") },
+            modifier = Modifier.fillMaxWidth()
+        )
+
+
+        //Spacer(modifier = Modifier.height(8.dp))
+
+
+// Username Input
+        var username by remember { mutableStateOf(TextFieldValue("")) }
+        OutlinedTextField(
+            value = username,
+            onValueChange = { username = it },
+            label = { Text(text = "Username:") },
+            modifier = Modifier.fillMaxWidth()
+        )
+
+
+        //Spacer(modifier = Modifier.height(8.dp))
+
+
+// Password Input
         var password by remember { mutableStateOf(TextFieldValue("")) }
         var passwordVisible by remember { mutableStateOf(false) }
         OutlinedTextField(
             value = password,
             onValueChange = { password = it },
-            label = { Text(text = "New Password") },
+            label = { Text(text = "Password:") },
             visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
             modifier = Modifier.fillMaxWidth()
         )
 
-        Spacer(modifier = Modifier.height(8.dp))
 
+        //Spacer(modifier = Modifier.height(8.dp))
+
+
+// Username Input
         var confirm by remember { mutableStateOf(TextFieldValue("")) }
         OutlinedTextField(
             value = confirm,
             onValueChange = { confirm = it },
-            label = { Text(text = "Confirm New Password") },
+            label = { Text(text = "Confirm Password:") },
             modifier = Modifier.fillMaxWidth()
         )
 
-        Spacer(modifier = Modifier.height(35.dp))
 
-        // Send Verification Code Button
+        //Spacer(modifier = Modifier.height(32.dp))
+
+
+        // Sign up Button
         Button(
             onClick = {  },
             modifier = Modifier
@@ -99,15 +135,15 @@ fun ChangePasswordView() {
             shape = RoundedCornerShape(30.dp),
             colors = ButtonDefaults.buttonColors(Color(0xFFFFFFFF))
         ) {
-            Text(text = "Change Password", fontSize = 18.sp, color = Color(0xFFFFA500))
+            Text(text = "Sign Up", fontSize = 18.sp, color = Color(0xFFFFA500))
         }
 
-        Spacer(modifier = Modifier.height(50.dp))
+        //Spacer(modifier = Modifier.height(50.dp))
     }
 }
 
 @Preview(showBackground = true)
 @Composable
-fun PreviewChangePasswordView() {
-    ChangePasswordView()
+fun PreviewSignupView() {
+    SignupView(rememberNavController())
 }
